So the number of visible chat message is no longer manually set.
It's now based on the height of the listbox. When the max was manually set to say 10, it was really easy to overflow the listbox box and end up not showing messages. Now if there are multiple really long messages, you just get fewer visible messages until more come in and displace it. Also changed how everything was rendered. The entire box is cleared/re-rendered on every message because otherwise you'd just get message drawing over each other and end up with a mess. Also removed boost::signals crap again, and made Anchor a PROPERTY.