Implement 3 new UI elements. Client parts of Ember.
Finally implemented BlockElement, ImageElement and added TextBlockElement, which is used to render multiline blocks of text. Also add beginning implementation of Ember, which is the new backend twitch proxy and control service. Plus misc changes to other stuff in reaction to all the work done on native UI side. Ember is a separate project under the ui repo.