Add binding for not-root elements
Also add support for the margin property to the ImageElement. Home screen is now the Activation screen. Add loads more events to EmberService and make some of the EmberService vars properties for some reason. Make EmberService urls properties, that are set in the main app class. Change steamlink build script so it knows about test and not-test builds.