Add options attribute to layout files. Implement the rest of the ECS events.
The only supported option ATM for the opts atribute is "nocache" to prevent LM from caching instances of a layout. Add QUALITYCHANGE, DISCONNECT and CHATUICHANGE ECS events. The QC event is for switching the quality of an already running stream, CHATUICHANGE will be for hiding/moving the chat, and DISCONNECT signals that the steam link receiving it is no longer connected to a user. Add event handler for EmberStreamStart event to Player.